expose useSync hook to outside components and add remaining time
